src/main/java/com/assignment/openweatherservice/api/WeatherController.java
package com.assignment.openweatherservice.api;

import com.assignment.openweatherservice.service.WeatherService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/weather")
public class WeatherController {

    private final WeatherService weatherService;

    public WeatherController(WeatherService weatherService) {
        this.weatherService = weatherService;
    }

    @GetMapping("/city/{city}")
    public WeatherResponse getByCity(@PathVariable String city) {
        return weatherService.getByCity(city);
    }
}
